requirements.txt
fastapi
uvicorn
pandas
numpy
sentence-transformers
faiss-cpu
streamlit
requests
beautifulsoup4
